# docs/openapi.yaml
openapi: 3.1.0
info:
  title: klaxond API
  version: 0.14.26
  summary: Notification bridge backend API
  description: |
    klaxond receives monitoring webhooks, renders notification payloads, manages
    routing/inhibition policy, and exposes a zero-build admin UI.

    Browser sessions use the `klaxond_session` cookie plus `X-Klaxond-CSRF` on
    same-origin mutations. Scoped machine clients should use Bearer tokens.
    Sensitive browser mutations can also return `428 Precondition Required`
    with `X-Klaxond-Reauth: required`; call `/api/auth/reauth` or complete passkey
    reauthentication before retrying.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
...
paths:
...
  /api/auth/login:
    get:
      tags: [Auth]
      operationId: loginPage
      summary: Render the public login page or redirect an existing session.
      security: []
      parameters:
        - name: return_to
          in: query
          schema:
            type: string
            default: /status
        - name: start
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: Starts the OIDC redirect when OIDC mode is active.
      responses:
        "200":
          description: Login HTML.
          content:
            text/html:
              schema:
                type: string
        "302":
          description: Redirect to OIDC or an already authenticated return URL.
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/auth/local/login:
    post:
      tags: [Auth]
      operationId: localLogin
      summary: Create a local or LDAP username/password session.
      description: Uses Argon2id local credentials in `basic` mode and LDAP bind verification in `ldap` mode.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JSON login response for fetch requests.
          headers:
            Set-Cookie:
              description: Signed `klaxond_session` cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "302":
          description: HTML form login redirects to the sanitized return URL.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/auth/magic/request:
    post:
      tags: [Auth]
      operationId: requestMagicLink
      summary: Request a short-lived local magic login link.
      description: >
        Returns `sent: true` for known and unknown users; when Klaxond can
        surface the link directly for the configured local user, the JSON
        response includes `link`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                return_to:
                  type: string
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
                return_to:
                  type: string
      responses:
        "200":
          description: Magic link request accepted.
          content:
            application/json:
              schema:
                type: object
                required: [sent, expiresInSeconds]
                properties:
                  sent:
                    type: boolean
                  link:
                    type:
                      - string
                      - "null"
                  expiresInSeconds:
                    type: integer
        "302":
          description: Browser form requests redirect to the callback when a local user matches, or back to the login page otherwise.
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/auth/callback:
    get:
      tags: [Auth]
      operationId: oidcCallback
      summary: Complete the OIDC authorization-code callback.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Session established and redirected to the original return URL.
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          description: OIDC discovery or token exchange failed.
  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Clear session cookies.
      security: []
      responses:
        "200":
          description: Session cookies cleared.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/auth/reauth:
    post:
      tags: [Auth]
      operationId: reauth
      summary: Refresh the short reauthentication window for local sessions.
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SudoRequest"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/SudoRequest"
      responses:
        "200":
          description: Reauthentication accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SudoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/auth/passkey/login/options:
    post:
      tags: [Auth]
      operationId: passkeyLoginOptions
      summary: Create passkey authentication options.
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  description: User subject, username or email for standalone passkey login.
                step_up:
                  type: string
                  description: One-time MFA / step-up token issued after primary authentication.
                return_to:
                  type: string
      responses:
        "200":
          description: WebAuthn authentication challenge.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/passkey/login/verify:
    post:
      tags: [Auth]
      operationId: passkeyLoginVerify
      summary: Verify passkey authentication and issue a session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Passkey session established.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/auth/step-up/passkey/register/options:
    post:
      tags: [Auth]
      operationId: stepUpPasskeyRegisterOptions
      summary: Create passkey registration options for an MFA / step-up challenge.
      description: Allows first-passkey enrollment only for the user bound to the pending step-up token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpPasskeyRegisterRequest"
      responses:
        "200":
          description: WebAuthn registration challenge.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/passkey/register/verify:
    post:
      tags: [Auth]
      operationId: stepUpPasskeyRegisterVerify
      summary: Verify a step-up passkey registration and issue the app session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Registered passkey plus issued session user.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/totp/setup/start:
    post:
      tags: [Auth]
      operationId: stepUpTotpSetupStart
      summary: Start TOTP enrollment for an MFA / step-up challenge.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpTokenRequest"
      responses:
        "200":
          description: TOTP setup payload scoped to the step-up user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepUpTotpSetup"
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/totp/setup/confirm:
    post:
      tags: [Auth]
      operationId: stepUpTotpSetupConfirm
      summary: Confirm TOTP enrollment and issue the app session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpTotpConfirmRequest"
      responses:
        "200":
          description: Issued session user and return URL.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/step-up/totp/verify:
    post:
      tags: [Auth]
      operationId: stepUpTotpVerify
      summary: Verify an existing TOTP factor and issue the app session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepUpTotpVerifyRequest"
      responses:
        "200":
          description: Issued session user and return URL.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
...
  /api/client-log:
    post:
      tags: [Operations]
      operationId: recordClientLog
      summary: Record a bounded frontend log entry in the backend log buffer.
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: string
                  enum: [info, warn, warning, error]
                key:
                  type: string
                message:
                  type: string
                path:
                  type: string
                stack:
                  type: string
                userAgent:
                  type: string
      responses:
        "204":
          description: Log entry accepted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          description: Payload is larger than the accepted frontend log limit.
...
  /api/auth/config:
    get:
      tags: [Auth]
      operationId: getAuthConfig
      summary: Return redacted auth settings and available auth modes/scopes.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Redacted auth settings.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Auth]
      operationId: updateAuthConfig
      summary: Replace auth settings after validation.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthConfigUpdate"
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/auth/tokens:
    get:
      tags: [Auth]
      operationId: listTokens
      summary: List API keys and PATs without token secrets.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Token metadata and available scopes.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Auth]
      operationId: createToken
      summary: Create a scoped API key or PAT.
      description: The raw token is returned once; store the prefix for later identification.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "200":
          description: Created token and one-time secret.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTokenResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/auth/totp/setup/start:
    post:
      tags: [Auth]
      operationId: startTotpSetup
      summary: Generate a local TOTP setup secret and otpauth URI.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      responses:
        "200":
          description: TOTP setup payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpSetup"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/auth/totp/setup/confirm:
    post:
      tags: [Auth]
      operationId: confirmTotpSetup
      summary: Enable local TOTP after validating the current code.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpEnableRequest"
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/auth/totp/disable:
    post:
      tags: [Auth]
      operationId: disableTotp
      summary: Disable local TOTP.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/auth/passkey/credentials:
    get:
      tags: [Auth]
      operationId: listPasskeys
      summary: List registered passkeys and public WebAuthn settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Passkey metadata.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/auth/passkey/register/options:
    post:
      tags: [Auth]
      operationId: passkeyRegisterOptions
      summary: Create passkey registration options for the current local user.
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "200":
          description: WebAuthn registration challenge.
          content:
            application/json:
              schema:
                type: object
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/auth/passkey/register/verify:
    post:
      tags: [Auth]
      operationId: passkeyRegisterVerify
      summary: Verify passkey registration.
      security:
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/config/import-preview:
    post:
      tags: [Config]
      operationId: previewConfigImport
      summary: Validate and diff a TOML or full settings bundle before restore.
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigBundle"
          text/plain:
            schema:
              type: string
      responses:
        "200":
          description: Import validation and diff result.
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/config/restore:
    post:
      tags: [Config]
      operationId: restoreConfig
      summary: Restore a TOML or full settings bundle after an automatic backup.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigBundle"
          text/plain:
            schema:
              type: string
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/channel-config:
    get:
      tags: [Routing]
      operationId: getChannelConfig
      summary: Return channel configuration and redacted secret presence/source flags.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Channel configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateChannelConfig
      summary: Update channel fields, including TOML-backed secrets when env overrides are not used.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/ntfy-topics:
    get:
      tags: [Routing]
      operationId: getNtfyTopics
      summary: Return ntfy topics, known severities and token presence flags.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: ntfy topic configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateNtfyTopics
      summary: Replace ntfy topic sidecar configuration.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/ingest-auth:
    get:
      tags: [Routing]
      operationId: getIngestAuth
      summary: Return webhook ingest authentication settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Ingest auth settings.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateIngestAuth
      summary: Update webhook ingest authentication settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/cascade-config:
    get:
      tags: [Routing]
      operationId: getCascadeConfig
      summary: Return cascade tier list and runtime/default flags.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Cascade configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Routing]
      operationId: updateCascadeConfig
      summary: Update cascade tiers and defaults.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/cascade/toggle:
    post:
      tags: [Routing]
      operationId: toggleCascade
      summary: Toggle or explicitly set the runtime cascade flag.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/dedup-config:
    get:
      tags: [Policy]
      operationId: getDedupConfig
      summary: Return dedup settings and pending queue counts.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Dedup configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Policy]
      operationId: updateDedupConfig
      summary: Replace dedup settings.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
  /api/inhibition-rules:
    get:
      tags: [Policy]
      operationId: listInhibitionRules
      summary: Return configured inhibition rules.
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200":
          description: Inhibition rule configuration.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
    post:
      tags: [Policy]
      operationId: updateInhibitionRules
      summary: Replace inhibition rules.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
  /api/inhibition-rules/test:
    post:
      tags: [Policy]
      operationId: testInhibitionRules
      summary: Test candidate inhibition rules without saving them.
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Rule test result.
          content:
            application/json:
              schema:
                type: object
        "4XX":
          $ref: "#/components/responses/ClientError"
  /api/inhibitions/clear:
    post:
      tags: [Policy]
      operationId: clearInhibitions
      summary: Clear active runtime inhibitions.
      security:
        - bearerAuth: []
        - sessionCookie: []
          csrfHeader: []
      responses:
        "200":
          $ref: "#/components/responses/OkJson"
        "428":
          $ref: "#/components/responses/ReauthRequired"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: klaxond scoped token
      description: API keys and PATs created from `/api/auth/tokens`.
    sessionCookie:
      type: apiKey
      in: cookie
      name: klaxond_session
      description: Signed browser session cookie.
    csrfHeader:
      type: apiKey
      in: header
      name: X-Klaxond-CSRF
      description: Required on browser-originated mutations when using a session cookie.
    basicAuth:
      type: http
      scheme: basic
      description: Username/password auth for non-browser clients in basic or LDAP mode.
...
  responses:
...
    ClientError:
      description: Generic client-side error, including invalid input, missing credentials or authorization failure depending on the active auth mode.
      content:
        text/plain:
          schema:
            type: string
...
  schemas:
...
    LoginResponse:
      type: object
      required: [ok, return_to, csrf]
      properties:
        ok:
          type: boolean
        return_to:
          type: string
        csrf:
          type: string
...
security:
  - bearerAuth: []
  - sessionCookie: []